@@ -6,29 +6,29 @@ publish_to: 'none'
 version: 0.0.1+1
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
-  flutter: '>=3.13.0'
+  sdk: '>=3.4.0 <4.0.0'
+  flutter: '>=3.22.0'
 
 dependencies:
-  dio: ^5.3.3
-  easy_localization: ^3.0.3
+  dio: ^5.4.3
+  easy_localization: ^3.0.7
   either_dart: ^1.0.0
   flutter:
     sdk: flutter
-  flutter_native_splash: ^2.3.5
-  flutter_svg: ^2.0.9
-  flutter_widget_from_html: ^0.14.7
-  get_it: ^7.6.4
-  go_router: ^12.1.1
-  injectable: ^2.3.2
-  url_launcher: ^6.2.1
+  flutter_native_splash: ^2.4.0
+  flutter_svg: ^2.0.10
+  flutter_widget_from_html: ^0.15.1
+  get_it: ^7.7.0
+  go_router: ^14.2.0
+  injectable: ^2.4.2
+  url_launcher: ^6.3.0
 
 dev_dependencies:
-  build_runner: ^2.4.6
-  flutter_lints: ^3.0.1
+  build_runner: ^2.4.11
+  flutter_lints: ^4.0.0
   flutter_test:
     sdk: flutter
-  injectable_generator: ^2.4.1
+  injectable_generator: ^2.6.1
 
 flutter:
   uses-material-design: true
